fix(daemon): harden local IPC QA findings
